test(storage): checkpoint acceptance suites and S0 contracts
Separate legacy Raft regressions, common acceptance scenarios and storage
module checks. Preserve bounded performance observations and their limits,
and record the storage redesign stages and S0 completion evidence.

Include the reviewed payload-size and table-page bounds fixes, checked
snapshot byte ranges, and range tests that retain valid checksums and
observe reads outside the granted slice without changing production APIs.

Keep result summaries and integrity manifests in Git while ignoring local
archives and raw runs. Remove generated caches and temporary builds.

Validation: 10 targeted C++ tests with ASan/UBSan/LSan, 25 harness contract
checks, targeted static checks, and 30 unchanged migrated C++/helper files.
Go is unavailable for a fresh model run; unchanged historical model and
E2E/performance evidence remain separately documented.

+++ src/include/catalog/column.h
@@ -95,7 +95,7 @@ class Column {
    * @param type type whose size is to be determined
    * @return size in bytes
    */
-  static auto TypeSize(TypeId type, uint32_t length = 0) -> uint8_t {
+  static auto TypeSize(TypeId type, uint32_t length = 0) -> uint32_t {
     switch (type) {
       case TypeId::BOOLEAN:
       case TypeId::TINYINT:
